Create Blinking led
Blinking LED

This code blinks an LED connected to pin 13 on and off every second.
It uses a simple loop to toggle the LED state and demonstrates the use of
digitalWrite() and delay() functions.

Hardware Requirements:
- Arduino board Uno
- LED
- Resistor (220 ohms or 1k ohm)

Note: This code uses the built-in LED on pin 13. If using an external LED, connect it to pin 13 and use a resistor to limit current.